requirements.txt
streamlit
yfinance
openai
finnhub-python
matplotlib
pandas
python-docx
